Refactor: feedback 반영
